Login fix: add whitelisted users to wss_user if they do not already exist
White listed users are those who are: already in wss_user, the primary or an
additional contact with a site.  This update fixes login for the latter two,
such that when a user logs in, if their number is white listed but they are
not in the wss_user table, then we create a record for them.

This update fixes an error where site-managers were unable to create a passcode.