Fix cancel confirmation request redirecting to an HTTP URL
JS code will already reload page after sending a cancel, so we do
not need to send a redirect back from the server. Worse yet,
this redirect back was causing an error as it was instructing
the web browser to go to an HTTP URL, which was triggered a front-end error.