Enable CA for wnc-supply-sites and sort the state list
Apply some hackery to enable CA.. (1) disable any
domain checks when fetching state, we don't need domain checks.
We will be consolidating to one domain that has all territories.
(2) Add CA as a deployment_state for wnc-supply-sites

Also in this update we sort the state list.. in the UI the list
of states is not sorted